Consolidate mocks
